# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PyWorld3 API",
    "description": "Run World3 what-if simulations",
    "version": "0.1.0"
  },
  "paths": {
    "/simulate": {
      "post": {
        "summary": "Simulate",
        "operationId": "simulate_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SimulationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/constants": {
      "get": {
        "summary": "Constants",
        "operationId": "constants_constants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "number"
                  },
                  "type": "object",
                  "title": "Response Constants Constants Get"
                }
              }
            }
          }
        }
      }
    },
    "/variables": {
      "get": {
        "summary": "Variables",
        "operationId": "variables_variables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Variables Variables Get"
                }
              }
            }
          }
        }
      }
    },
    "/presets": {
      "get": {
        "summary": "Get Presets",
        "description": "List available built-in scenario presets.",
        "operationId": "get_presets_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PresetInfo"
                  },
                  "type": "array",
                  "title": "Response Get Presets Presets Get"
                }
              }
            }
          }
        }
      }
    },
    "/simulate/preset/{preset_name}": {
      "post": {
        "summary": "Simulate Preset",
        "description": "Run a simulation using a built-in preset, with optional overrides.",
        "operationId": "simulate_preset_simulate_preset__preset_name__post",
        "parameters": [
          {
            "name": "preset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SimulationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Overrides"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/metadata/constants": {
      "get": {
        "summary": "Constant Metadata",
        "description": "Return metadata (full name, unit, sector) for every tunable constant.",
        "operationId": "constant_metadata_metadata_constants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response Constant Metadata Metadata Constants Get"
                }
              }
            }
          }
        }
      }
    },
    "/metadata/variables": {
      "get": {
        "summary": "Variable Metadata",
        "description": "Return metadata (full name, unit, sector) for every output variable.",
        "operationId": "variable_metadata_metadata_variables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response Variable Metadata Metadata Variables Get"
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "post": {
        "summary": "Compare",
        "description": "Compare two scenarios side by side.",
        "operationId": "compare_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompareMetric": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "variable": {
            "type": "string",
            "title": "Variable"
          },
          "value_a": {
            "type": "number",
            "title": "Value A"
          },
          "value_b": {
            "type": "number",
            "title": "Value B"
          },
          "delta_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Pct"
          }
        },
        "type": "object",
        "required": [
          "label",
          "variable",
          "value_a",
          "value_b",
          "delta_pct"
        ],
        "title": "CompareMetric"
      },
      "CompareRequest": {
        "properties": {
          "scenario_a": {
            "$ref": "#/components/schemas/ScenarioSpec"
          },
          "scenario_b": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "scenario_a"
        ],
        "title": "CompareRequest"
      },
      "CompareResponse": {
        "properties": {
          "scenario_a": {
            "type": "string",
            "title": "Scenario A"
          },
          "scenario_b": {
            "type": "string",
            "title": "Scenario B"
          },
          "results_a": {
            "$ref": "#/components/schemas/SimulationResponse"
          },
          "results_b": {
            "$ref": "#/components/schemas/SimulationResponse"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/CompareMetric"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "scenario_a",
          "scenario_b",
          "results_a",
          "results_b",
          "metrics"
        ],
        "title": "CompareResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PresetInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "constants": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Constants"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "constants"
        ],
        "title": "PresetInfo"
      },
      "ScenarioSpec": {
        "properties": {
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ScenarioSpec",
        "description": "One side of a comparison: either a preset name or an inline request."
      },
      "SimulationRequest": {
        "properties": {
          "year_min": {
            "type": "number",
            "maximum": 2500.0,
            "minimum": 1900.0,
            "title": "Year Min",
            "description": "Start year of the simulation",
            "default": 1900
          },
          "year_max": {
            "type": "number",
            "maximum": 2500.0,
            "minimum": 1900.0,
            "title": "Year Max",
            "description": "End year of the simulation",
            "default": 2100
          },
          "dt": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Dt",
            "description": "Time step of the simulation [year]",
            "default": 0.5
          },
          "pyear": {
            "type": "number",
            "title": "Pyear",
            "description": "Implementation date of new policies [year]",
            "default": 1975
          },
          "iphst": {
            "type": "number",
            "title": "Iphst",
            "description": "Implementation date of new policy on health service time [year]",
            "default": 1940
          },
          "constants": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constants",
            "description": "Override any World3 constant by name (e.g. {'nri': 2e12})"
          },
          "output_variables": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Variables",
            "description": "Which output variables to return (default: a curated list)"
          }
        },
        "type": "object",
        "title": "SimulationRequest"
      },
      "SimulationResponse": {
        "properties": {
          "year_min": {
            "type": "number",
            "title": "Year Min"
          },
          "year_max": {
            "type": "number",
            "title": "Year Max"
          },
          "dt": {
            "type": "number",
            "title": "Dt"
          },
          "time": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Time"
          },
          "constants_used": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Constants Used"
          },
          "series": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TimeSeriesOutput"
            },
            "type": "object",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "year_min",
          "year_max",
          "dt",
          "time",
          "constants_used",
          "series"
        ],
        "title": "SimulationResponse"
      },
      "TimeSeriesOutput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "name",
          "values"
        ],
        "title": "TimeSeriesOutput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}